back-end/Wirelesworks/Scan/requirements.txt
alive-progress==3.1.5
termcolor==2.3.0
pyserial==3.5
pyyaml==6.0.1
websockets==12.0